requirements.txt
absl-py==2.0.0
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.3
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.1.0
blinker==1.7.0
boto3==1.34.65
botocore==1.34.65
bottle==0.12.19
CacheControl==0.12.6
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
cfractions==2.2.0
channels==4.0.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.3
commonmark==0.9.1
constantly==23.10.4
contextlib2==0.6.0
coverage==7.4.4
crispy-bootstrap5==2024.2
cryptography==41.0.7
cupy-cuda12x==12.0.0
cx-Freeze==6.15.9
cx-Logging==3.1.0
cycler==0.11.0
daphne==4.1.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.0
distro==1.9.0
dj-database-url==0.5.0
Django==4.2.11
django-appconf==1.0.6
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-compressor==4.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-environ==0.11.2
django-except-catcher==1.0.1
django-filter==24.1
django-heroku==0.3.1
django-import-export==3.3.7
django-js-asset==2.2.0
django-json-widget==1.1.1
django-mathfilters==1.0.0
django-model-utils==4.4.0
django-multiselectfield==0.1.13
django-on-heroku==1.1.2
django-property-filter==1.2.0
django-sass-processor==1.4
django-simple-history==3.5.0
django-storages==1.14.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.4
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastrlock==0.8.1
filelock==3.6.0
firebase-admin==5.1.0
Flask==2.0.2
flatbuffers==23.5.26
fonttools==4.31.2
fpdf==1.7.2
future==1.0.0
gast==0.4.0
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.23.4
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.59.3
grpcio-status==1.42.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
html5lib==1.0.1
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
image==1.5.33
importlib_metadata==7.0.2
incremental==22.10.0
iniconfig==2.0.0
ipaddr==2.2.0
iso8601==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
jiter==0.6.1
jmespath==1.0.1
jws==0.1.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.1
libclang==16.0.6
libsass==0.23.0
lief==0.13.2
lockfile==0.12.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.5.1
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.8
node==1.0
nplusone==1.0.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
odict==1.9.0
openai==1.51.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pdfkit==1.0.0
pefile==2023.2.7
pep517==0.8.2
pillow==10.2.0
pipenv==2022.4.21
platformdirs==2.5.2
pluggy==1.4.0
plumber==1.7
progress==1.5
proto-plus==1.19.8
protobuf==3.19.6
psycopg2==2.9.9
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.5.2
Pygments==2.17.2
pyinstaller==5.13.1
pyinstaller-hooks-contrib==2023.7
PyJWT==2.8.0
pymongo==3.11.3
pyngrok==5.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyreadline3==3.4.1
Pyrebase==3.0.27
Pyrebase4==4.5.0
pyserial==3.5
pyTelegramBotAPI==4.2.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-jwt==2.0.1
pytoml==0.1.21
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.3
rich==13.7.1
rjsmin==1.2.1
rsa==4.8
s3transfer==0.10.1
sendgrid==6.11.0
sentry-sdk==1.42.0
serial==0.0.97
service-identity==24.1.0
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tablib==3.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.5
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==1.4.0
xlwt==1.3.0
zipp==3.18.1
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==6.2
zope.lifecycleevent==4.4
zope.proxy==4.5.0
